Update to the IsCyrillic Function
The updated 'IsCyrillic' function now takes into account the exclusion of characters listed in the 'ignoredCharacters' list. This ensures an accurate determination of whether a character is Cyrillic and excludes ignored alternative characters from being considered as Russian characters. Characters from the 'ignoredCharacter's list will no longer be counted as Russian characters.

The updated code also includes the 'tableContains' function, which is used to check whether a character is included in the list of ignored characters.